+++ Project_SWENG/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -72,6 +72,7 @@ MonoBehaviour:
   - ReadyCancleToServer
   - ReadyToServer
   - PlayerLeave
+  - ReadyStateChange
   DisableAutoOpenWizard: 1
   ShowSettings: 0
   DevRegionSetOnce: 1
