+++ Project_SWENG/Assets/Scripts/GameManagers/NetworkManager.Chatting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15c81048d5754584b966f0c544763164
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
