+++ Project_SWENG/Assets/Scripts/GameManagers/NetworkManager.Room.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e17377bb695ea36498726099906638e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
